fix(ci): address review action feedback
